+++ lib/services/wishlist_service.dart
@@ -1,9 +1,7 @@
 import 'package:MobileOne/dao/wishlist_dao.dart';
-import 'package:MobileOne/data/categories.dart';
 import 'package:MobileOne/data/owner_details.dart';
 import 'package:MobileOne/data/wishlist.dart';
 import 'package:MobileOne/data/wishlist_item.dart';
-import 'package:MobileOne/localization/localization.dart';
 import 'package:MobileOne/services/analytics_services.dart';
 import 'package:MobileOne/services/image_service.dart';
 import 'package:MobileOne/services/messaging_service.dart';
@@ -26,7 +24,6 @@ class WishlistService {
   Map<String, dynamic> _ownerLists = {};
   Map<String, dynamic> _guestLists = {};
   Map<String, OwnerDetails> _ownerDetails = {};
-  List<Categories> _categories = [];
 
   void _flush() {
     _wishlists = {};
@@ -35,7 +32,6 @@ class WishlistService {
     _ownerLists = {};
     _guestLists = {};
     _ownerDetails = {};
-    _categories = [];
   }
 
   List get ownerLists => _ownerLists[userService.user.uid];
@@ -290,31 +286,6 @@ class WishlistService {
     }
   }
 
-  List<Categories> getCategories() => _categories;
-
-  Future<List<Categories>> fetchCategories(BuildContext context) async {
-    _categories
-        .add(Categories.fromMap(null, getString(context, "null_category")));
-    List<Map<String, dynamic>> map = await dao.getCategories();
-    map.forEach((element) {
-      switch (element["id"]) {
-        case "78amWnyUJe3ekEs9FD53":
-          _categories.add(Categories.fromMap(
-              element["id"], getString(context, "recipy_category")));
-          break;
-        case "8jzs8g05JvvVQ87DI9wL":
-          _categories.add(Categories.fromMap(
-              element["id"], getString(context, "food_category")));
-          break;
-        default:
-          _categories.add(Categories.fromMap(element["id"], element["label"]));
-          break;
-      }
-    });
-
-    return _categories;
-  }
-
   changeWishlistCategory(String wishlistUuid, String categoryId) async {
     dao.changeWishlistCategory(wishlistUuid, categoryId);
     _wishlists[wishlistUuid].categoryId = categoryId;
@@ -394,5 +365,15 @@ class WishlistService {
       _wishlists[wishlistUuid].lastModification =
           Timestamp.fromDate(DateTime.now());
     }
+  }
+
+  bool isCategoryDeletable(String categoryUuid) {
+    bool isCategoryDeletable = true;
+    _wishlists.values.forEach((element) {
+      if (element.categoryId == categoryUuid) {
+        isCategoryDeletable = false;
+      }
+    });
+    return isCategoryDeletable;
   }
 }