+++ CHANGELOG.md
@@ -48,6 +48,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add new units for items
 - Add sort wishlists
 - Add tutorial
+- Add category management
 
 ### Changed